各種調整
keymap.cに書いていたロジックをok60_hhkb.cに移動
キーマップはdefaultに統合(もともとidle_timerと差異がなかった)
アイドルタイマーの有効無効設定をrules.mkに追加
