+++ README.md
@@ -4,16 +4,18 @@
 
 ## OK60 HHKB
 
-* default
-    - Caps lock key assigned to MO(1) + Tab key
+    * Caps lock key assigned to MO(1) + Tab key
+    * Turn the backlight off automatically in 30 seconds.
+    * Turn LEDs off automatically in 5 minutes.
 
     `make ok60_hhkb:default`
 
-* idle_timer
-    - Turn the backlight off automatically in 30 seconds.
-    - Turn LEDs off automatically in 5 minutes.
+    * If you do not need an idle timer, comment out IDLE_TIMER_ENABLE in rules.mk as shown below.
 
-    `make ok60_hhkb:idle_timer`
+        * Backlight and LEDs idle timer is disable.
+            ```
+            #IDLE_TIMER_ENABLE = yes
+            ```
 
 ## Keymap
